Fixed name of BootStrap.groovy so it is included as part of the plugin, updated version number for the plugin, added null check in the filter and moved the annotation out of the controller package (It never belonged there)
