+++ .settings/com.springsource.sts.grails.core.prefs
@@ -1,4 +1,4 @@
-#Sun Dec 18 08:32:22 EST 2011
-com.springsource.sts.grails.core.com.springsource.sts.grails.core.install.name=Grails 1.3.7
-com.springsource.sts.grails.core.use.default.install=false
+#Thu Feb 23 11:56:46 CST 2012
+com.springsource.sts.grails.core.com.springsource.sts.grails.core.install.name=Grails 2.0
+com.springsource.sts.grails.core.use.default.install=true
 eclipse.preferences.version=1
