+++ .settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,4 @@
-#Sun Dec 18 08:32:30 EST 2011
+#Thu Feb 23 11:56:27 CST 2012
 eclipse.preferences.version=1
 encoding//grails-app/i18n=UTF-8
+encoding//grails-app/i18n/messages.properties=UTF-8
