+++ application.properties
@@ -1,5 +1,6 @@
 #Grails Metadata file
-#Mon Jan 30 20:52:50 CST 2012
+#Thu Feb 23 14:32:06 CST 2012
 app.grails.version=2.0.0
 app.name=grails-feature-toggle
+plugins.svn=1.0.1
 plugins.tomcat=2.0.0
