+++ grails-app/conf/grails/feature/toggle/filters/FeatureToggleFilters.groovy
@@ -1,6 +1,5 @@
 package grails.feature.toggle.filters
 import grails.plugin.featuretoggle.FeatureToggleService;
-import grails.plugin.featuretoggle.annotations.Feature;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 
@@ -8,36 +7,38 @@ class FeatureToggleFilters implements ApplicationContextAware {
 
 	FeatureToggleService featureToggleService
 	ApplicationContext applicationContext
-	
+
 	def filters = {
 		allControllers(controller:'*', action:'*') {
-			before = { 
+			before = {
 				def controllers = request.servletContext['controlledActions']
-				
-				def curAction = actionName
 
-				// In the case of the action not being specified (default action) we need to look up 
-				// if it was configured in the controller or it is using the convention 'index'
-				if(curAction == null) {
-					def artefact = grailsApplication.getArtefactByLogicalPropertyName("Controller", controllerName)
+				if(controllers != null) {
+					def curAction = actionName
 
-					def bean = applicationContext.getBean(artefact.clazz.name)
-					def metaProperty = bean.metaClass.getMetaProperty("defaultAction")
+					// In the case of the action not being specified (default action) we need to look up
+					// if it was configured in the controller or it is using the convention 'index'
+					if(curAction == null) {
+						def artefact = grailsApplication.getArtefactByLogicalPropertyName("Controller", controllerName)
 
-					if(metaProperty) {
-						curAction = bean.defaultAction
-					} else {
-						curAction = 'index'
+						def bean = applicationContext.getBean(artefact.clazz.name)
+						def metaProperty = bean.metaClass.getMetaProperty("defaultAction")
+
+						if(metaProperty) {
+							curAction = bean.defaultAction
+						} else {
+							curAction = 'index'
+						}
 					}
-				}
-				
-				def action = controllers[controllerName?.toLowerCase() + '.' + curAction?.toLowerCase()] 
-				
-				if(action != null && !featureToggleService.isFeatureEnabled(action.name)) {
-					if(action.resultRedirect.size() > 0) {
-						redirect(uri: action.resultRedirect)
-					} else {
-						render(status: action.resultStatus)
+
+					def action = controllers[controllerName?.toLowerCase() + '.' + curAction?.toLowerCase()]
+
+					if(action != null && !featureToggleService.isFeatureEnabled(action.name)) {
+						if(action.resultRedirect.size() > 0) {
+							redirect(uri: action.resultRedirect)
+						} else {
+							render(status: action.resultStatus)
+						}
 					}
 				}
 			}
