+++ grails-app/controllers/grails/feature/toggle/SampleController.groovy
@@ -1,6 +1,6 @@
 package grails.feature.toggle
 
-import grails.plugin.featuretoggle.annotations.Feature;
+import grails.plugin.feature.toggle.annotations.Feature
 
 @Feature(name='sample')
 class SampleController {
