+++ src/groovy/grails/plugin/feature/toggle/annotations/Feature.java
@@ -1,4 +1,4 @@
-package grails.plugin.featuretoggle.annotations;
+package grails.plugin.feature.toggle.annotations;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
